JAMES-4124 Extract contact tests for redis event bus
avoid duplicating too much test code.
